refactor(simulator): move internal from cmd
