+++ simulator/internal/patient.go
