+++ simulator/internal/state.go
