+++ simulator/internal/vitals.go
